+++ example-scripts/meta-edge/install-new-identity-without-parsec.sh
@@ -1,4 +1,8 @@
+#!/bin/bash
+#
 # Copyright (c) 2020, Arm Limited and affiliates.
+# Copyright (c) 2023, Izuma Networks
+#
 # SPDX-License-Identifier: Apache-2.0
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -23,12 +27,12 @@ PDM_CRED_DIR="/userdata/mbed/mcc_config"
 FCC_OUTPUT="pal/"
 
 # Move the identity.json to location as per maestro config
-mkdir -p $EDGE_GW_CONFIG_DIR
-mv $IDENTITY_JSON_FILE $EDGE_GW_CONFIG_DIR
+mkdir -p "$EDGE_GW_CONFIG_DIR"
+mv "$IDENTITY_JSON_FILE" "$EDGE_GW_CONFIG_DIR"
 echo "Successfully installed identity.json at $EDGE_GW_CONFIG_DIR$IDENTITY_JSON_FILE"
 
 # Move the device management credentials to location as per edge-core config
-mv $FCC_OUTPUT $PDM_CRED_DIR
+mv "$FCC_OUTPUT" "$PDM_CRED_DIR"
 echo "Successfully installed device management credentials at $PDM_CRED_DIR"
 
 echo "Done"